send inv computed tasks *after* commit to avoid a race condition
